+++ src/lib.rs
@@ -221,7 +221,9 @@ pub use builder::*;
 pub use builder_dissect::dissect_builder_xml;
 pub use event_loops_bridge::{block_on, run_actix_inside_gtk_event_loop, try_block_on};
 pub use signal::{Signal, SignalResult};
-pub use signal_routing::{route_signal, GenerateRoutingGtkHandler, IntoGenerateRoutingGtkHandler, RawSignalCallback};
+pub use signal_routing::{
+    route_signal, GenerateRoutingGtkHandler, IntoGenerateRoutingGtkHandler, NamespacedSignalRouter, RawSignalCallback,
+};
 
 #[derive(thiserror::Error, Debug)]
 pub enum Error {